Reorganize tests for validateTalent
